+++ goagen/gen_controller/generator_test.go
@@ -1,62 +1,11 @@
 package gencontroller_test
 
 import (
-	"io/ioutil"
-	"os"
-	"path/filepath"
-	"strings"
-
 	"github.com/goadesign/goa/design"
 	"github.com/goadesign/goa/goagen/gen_controller"
-	"github.com/goadesign/goa/version"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"github.com/onsi/gomega/gexec"
 )
-
-var _ = Describe("Generate", func() {
-	const testgenPackagePath = "github.com/goadesign/goa/goagen/gen_controller/goatest"
-
-	var outDir string
-	var files []string
-	var genErr error
-
-	BeforeEach(func() {
-		gopath := filepath.SplitList(os.Getenv("GOPATH"))[0]
-		outDir = filepath.Join(gopath, "src", testgenPackagePath)
-		err := os.MkdirAll(outDir, 0777)
-		Ω(err).ShouldNot(HaveOccurred())
-		os.Args = []string{"goagen", "--out=" + outDir, "--design=foo", "--version=" + version.String()}
-	})
-
-	JustBeforeEach(func() {
-		files, genErr = gencontroller.Generate()
-	})
-
-	AfterEach(func() {
-		os.RemoveAll(outDir)
-	})
-
-	Context("with a dummy API", func() {
-		BeforeEach(func() {
-			design.Design = &design.APIDefinition{
-				Name:        "test api",
-				Title:       "dummy API with no resource",
-				Description: "I told you it's dummy",
-			}
-		})
-
-		It("generates a dummy app", func() {
-			Ω(genErr).Should(BeNil())
-			Ω(files).Should(HaveLen(1))
-			content, err := ioutil.ReadFile(filepath.Join(outDir, "controller.go"))
-			Ω(err).ShouldNot(HaveOccurred())
-			Ω(len(strings.Split(string(content), "\n"))).Should(BeNumerically(">=", 16))
-			_, err = gexec.Build(testgenPackagePath)
-			Ω(err).ShouldNot(HaveOccurred())
-		})
-	})
-})
 
 var _ = Describe("NewGenerator", func() {
 	var generator *gencontroller.Generator
