+++ goagen/gen_main/generator.go
@@ -2,12 +2,10 @@ package genmain
 
 import (
 	"flag"
-	"fmt"
 	"net"
 	"os"
 	"path"
 	"path/filepath"
-	"strings"
 	"text/template"
 
 	"github.com/goadesign/goa/design"
@@ -82,60 +80,24 @@ func (g *Generator) Generate() (_ []string, err error) {
 	if g.Force {
 		os.Remove(mainFile)
 	}
-	funcs := template.FuncMap{
-		"tempvar":   tempvar,
-		"okResp":    g.okResp,
-		"targetPkg": func() string { return g.Target },
-	}
-	imp, err := codegen.PackagePath(g.OutDir)
-	if err != nil {
-		return nil, err
-	}
-	imp = path.Join(filepath.ToSlash(imp), "app")
 	_, err = os.Stat(mainFile)
 	if err != nil {
 		// ensure that the output directory exists before creating a new main
 		if err := os.MkdirAll(g.OutDir, 0755); err != nil {
 			return nil, err
 		}
-		if err = g.createMainFile(mainFile, funcs); err != nil {
+		if err = g.createMainFile(mainFile, utils.GetFunctions(g.Target)); err != nil {
 			return nil, err
 		}
 	}
-	imports := []*codegen.ImportSpec{
-		codegen.SimpleImport("io"),
-		codegen.SimpleImport("github.com/goadesign/goa"),
-		codegen.SimpleImport(imp),
-		codegen.SimpleImport("golang.org/x/net/websocket"),
-	}
+
 	err = g.API.IterateResources(func(r *design.ResourceDefinition) error {
-		filename := filepath.Join(g.OutDir, codegen.SnakeCase(r.Name)+".go")
-		if g.Force {
-			os.Remove(filename)
+		filename, err := utils.GenerateControllerFile(g.Force, g.Target, g.OutDir, "main", r.Name, r)
+		if err != nil {
+			return err
 		}
-		if _, e := os.Stat(filename); e != nil {
-			g.genfiles = append(g.genfiles, filename)
-			file, err2 := codegen.SourceFileFor(filename)
-			if err2 != nil {
-				return err
-			}
-			file.WriteHeader("", "main", imports)
-			if err2 = file.ExecuteTemplate("controller", ctrlT, funcs, r); err2 != nil {
-				return err
-			}
-			err2 = r.IterateActions(func(a *design.ActionDefinition) error {
-				if a.WebSocket() {
-					return file.ExecuteTemplate("actionWS", actionWST, funcs, a)
-				}
-				return file.ExecuteTemplate("action", actionT, funcs, a)
-			})
-			if err2 != nil {
-				return err
-			}
-			if err2 = file.FormatCode(); err2 != nil {
-				return err2
-			}
-		}
+
+		g.genfiles = append(g.genfiles, filename)
 		return nil
 	})
 	if err != nil {
@@ -151,18 +113,6 @@ func (g *Generator) Cleanup() {
 		os.Remove(f)
 	}
 	g.genfiles = nil
-}
-
-// tempCount is the counter used to create unique temporary variable names.
-var tempCount int
-
-// tempvar generates a unique temp var name.
-func tempvar() string {
-	tempCount++
-	if tempCount == 1 {
-		return "c"
-	}
-	return fmt.Sprintf("c%d", tempCount)
 }
 
 func (g *Generator) createMainFile(mainFile string, funcs template.FuncMap) error {
@@ -201,57 +151,6 @@ func (g *Generator) createMainFile(mainFile string, funcs template.FuncMap) erro
 	return file.FormatCode()
 }
 
-func (g *Generator) okResp(a *design.ActionDefinition) map[string]interface{} {
-	var ok *design.ResponseDefinition
-	for _, resp := range a.Responses {
-		if resp.Status == 200 {
-			ok = resp
-			break
-		}
-	}
-	if ok == nil {
-		return nil
-	}
-	var mt *design.MediaTypeDefinition
-	var ok2 bool
-	if mt, ok2 = design.Design.MediaTypes[design.CanonicalIdentifier(ok.MediaType)]; !ok2 {
-		return nil
-	}
-	view := ok.ViewName
-	if view == "" {
-		view = design.DefaultView
-	}
-	pmt, _, err := mt.Project(view)
-	if err != nil {
-		return nil
-	}
-	var typeref string
-	if pmt.IsError() {
-		typeref = `goa.ErrInternal("not implemented")`
-	} else {
-		name := codegen.GoTypeRef(pmt, pmt.AllRequired(), 1, false)
-		var pointer string
-		if strings.HasPrefix(name, "*") {
-			name = name[1:]
-			pointer = "*"
-		}
-		typeref = fmt.Sprintf("%s%s.%s", pointer, g.Target, name)
-		if strings.HasPrefix(typeref, "*") {
-			typeref = "&" + typeref[1:]
-		}
-		typeref += "{}"
-	}
-	var nameSuffix string
-	if view != "default" {
-		nameSuffix = codegen.Goify(view, true)
-	}
-	return map[string]interface{}{
-		"Name":    ok.Name + nameSuffix,
-		"GoType":  codegen.GoNativeType(pmt),
-		"TypeRef": typeref,
-	}
-}
-
 const mainT = `
 func main() {
 	// Create service
@@ -271,50 +170,6 @@ func main() {
 	// Start service
 	if err := service.ListenAndServe(":{{ getPort .API.Host }}"); err != nil {
 		service.LogError("startup", "err", err)
-	}
-}
-`
-
-const ctrlT = `// {{ $ctrlName := printf "%s%s" (goify .Name true) "Controller" }}{{ $ctrlName }} implements the {{ .Name }} resource.
-type {{ $ctrlName }} struct {
-	*goa.Controller
-}
-
-// New{{ $ctrlName }} creates a {{ .Name }} controller.
-func New{{ $ctrlName }}(service *goa.Service) *{{ $ctrlName }} {
-	return &{{ $ctrlName }}{Controller: service.NewController("{{ $ctrlName }}")}
-}
-`
-
-const actionT = `{{ $ctrlName := printf "%s%s" (goify .Parent.Name true) "Controller" }}// {{ goify .Name true }} runs the {{ .Name }} action.
-func (c *{{ $ctrlName }}) {{ goify .Name true }}(ctx *{{ targetPkg }}.{{ goify .Name true }}{{ goify .Parent.Name true }}Context) error {
-	// {{ $ctrlName }}_{{ goify .Name true }}: start_implement
-
-	// Put your logic here
-
-	// {{ $ctrlName }}_{{ goify .Name true }}: end_implement
-{{ $ok := okResp . }}{{ if $ok }} res := {{ $ok.TypeRef }}
-{{ end }} return {{ if $ok }}ctx.{{ $ok.Name }}(res){{ else }}nil{{ end }}
-}
-`
-
-const actionWST = `{{ $ctrlName := printf "%s%s" (goify .Parent.Name true) "Controller" }}// {{ goify .Name true }} runs the {{ .Name }} action.
-func (c *{{ $ctrlName }}) {{ goify .Name true }}(ctx *{{ targetPkg }}.{{ goify .Name true }}{{ goify .Parent.Name true }}Context) error {
-	c.{{ goify .Name true }}WSHandler(ctx).ServeHTTP(ctx.ResponseWriter, ctx.Request)
-	return nil
-}
-
-// {{ goify .Name true }}WSHandler establishes a websocket connection to run the {{ .Name }} action.
-func (c *{{ $ctrlName }}) {{ goify .Name true }}WSHandler(ctx *{{ targetPkg }}.{{ goify .Name true }}{{ goify .Parent.Name true }}Context) websocket.Handler {
-	return func(ws *websocket.Conn) {
-		// {{ $ctrlName }}_{{ goify .Name true }}: start_implement
-
-		// Put your logic here
-
-		// {{ $ctrlName }}_{{ goify .Name true }}: end_implement
-		ws.Write([]byte("{{ .Name }} {{ .Parent.Name }}"))
-		// Dummy echo websocket server
-		io.Copy(ws, ws)
 	}
 }
 `
